Fix broken tests
It seems some changes to the SDK cause the cell to no longer be setup on dequeue. This manually adds them to the view hierarchy to have the recursive walk up work. Additionally, the RecordingTableView will return the index path for a cell override; this was needed to mock out the index path finding portion of the scheme locator.